Initial Upload to Github
First version of code was completely copy and paste the same thing over and over again to add new sounds. Since the bot was built quickly in order for it to be relevant at the time this was okay, unfortunately it was very 'fragile' and as more sounds were added the code became unneccessarily long. This version is finally presentable and efficient adding a queue, proper login support and easy to add new sounds.